Добавлена рабочая функция удаления записи

+++ src/main/java/BDConnection/StudentService.java
@@ -8,13 +8,13 @@ import java.util.List;
  * Created by DrSwitch on 20.04.2017.
  */
 public interface StudentService {
-    StudentEntity getByIdStudent(long idStudent);
+    StudentEntity getByIdStudent(int idStudent);
 
     StudentEntity addStudent(StudentEntity student);
 
     void updateStudent(StudentEntity student);
 
-    void deleteStudent(long idStudent);
+    void deleteStudent(int idStudent);
 
     List<StudentEntity> getAllStudent();
 }
